requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==5.0.6
django-cors-headers==4.3.1
django-daraja==1.3.0
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
redis==5.0.4
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
